refactor: 인기 테마 조회 limit을 서비스 상수에서 쿼리 파라미터로 변경

+++ src/main/java/roomescape/repository/ThemeRepository.java
@@ -60,7 +60,7 @@ public class ThemeRepository {
         return jdbcTemplate.query(sql, themeRowMapper);
     }
 
-    public List<Theme> findByCurrentDateAndLastWeekDateAndLimit(String currentDate, String lastWeekDate, int limit) {
+    public List<Theme> findByCurrentDateAndLastWeekDateAndLimit(String currentDate, String lastWeekDate, long limit) {
         String sql = """
                 SELECT
                     t.id,
